~ Relocated GameHook-files ~ Introduces game list

+++ 0_vs/thearcadearcade/Source/GameHooks/Emulator.cs
@@ -242,8 +242,11 @@ namespace thearcadearcade.GameHooks
         }
 
         ~Emulator()
-        {
-            CurrentProcess.Kill();
+        { 
+            if (!CurrentProcess.HasExited)
+            {
+               CurrentProcess.Kill();
+            }
         }
     }
 }
